backend/requirements.txt

# Professional Project Requirements

# Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0

# Text Analysis Libraries
nltk==3.8.1
textstat==0.7.3

# File Processing Libraries  
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.4